some temp cleanup
